Order Beans
Created the 'Order' beans and added many JPA annotations, some of which
will need testing once combined with the 'Client' beans. A change was
made to the database schema: PRODUCT_IMAGE(BLOB) was replaced with
PRODUCT_IMAGE_ID(NUMBER) and a corresponding table was added
IMS_PRODUCT_IMAGE{IMAGE_ID NUMBER, IMAGE_ADDRESS, VARCHAR2(2000)}. This
table has been made to hold the address of images on the S3 server and
potentially to hold extra information (like the caption/description for
images that failed to load) in the future.